src/requirements.txt
boto3
cffi
pynacl
requests
